+++ src/Config/IniOverrideConfig.php
@@ -40,4 +40,15 @@ class IniOverrideConfig
     {
         ini_set('memory_limit', $this->getDefaultMemoryLimit());
     }
+
+    /**
+     * Call this function for all PHP script running from the CLI (ex: all symfony cron commands) when you are on CleverCloud to ensure every date
+     * creation or comparaison are using the right timezone.
+     * Alternativaly you can also set it directly when invoking the PHP command using the -d option like so :
+     *  php -d date.timezone="Europe/Paris" bin/console app:my-command
+     */
+    public function initDefaultTimezoneForCli(string $timezone = 'Europe/Paris'): void
+    {
+        date_default_timezone_set($timezone);
+    }
 }