+++ src/Entity/EmailInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Smart\CoreBundle\Entity;
+
+interface EmailInterface
+{
+    public function getEmail(): ?string;
+
+    public function setEmail(?string $email): self;
+}