+++ src/Entity/NameableInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Smart\CoreBundle\Entity;
+
+interface NameableInterface
+{
+    public function getName(): ?string;
+
+    public function setName(string $name): self;
+}