+++ src/Entity/SirenInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Smart\CoreBundle\Entity;
+
+interface SirenInterface
+{
+    public function getSiren(): ?string;
+
+    public function setSiren(?string $siren): static;
+}