+++ src/Monitoring/ProcessMonitor.php
@@ -1,0 +1,42 @@
+<?php
+
+namespace Smart\CoreBundle\Monitoring;
+
+use Doctrine\ORM\EntityManagerInterface;
+use Smart\CoreBundle\Entity\ProcessInterface;
+use Smart\CoreBundle\Enum\ProcessStatusEnum;
+
+/**
+ * @author Mathieu Ducrot <mathieu.ducrot@smartbooster.io>
+ */
+class ProcessMonitor
+{
+    public function __construct(private readonly EntityManagerInterface $entityManager)
+    {
+    }
+
+    public function start(ProcessInterface $process): ProcessInterface
+    {
+        $process->setStartedAt(new \DateTime());
+        $process->setStatus(ProcessStatusEnum::ONGOING);
+
+        return $process;
+    }
+
+    public function end(ProcessInterface $process, bool $isSuccess = true, bool $flush = true): void
+    {
+        $endedAt = new \DateTime();
+        $process->setEndedAt($endedAt);
+        $process->setDuration($endedAt->getTimestamp() - $process->getStartedAt()->getTimestamp());
+        if ($isSuccess) {
+            $process->setStatus(ProcessStatusEnum::SUCCESS);
+        } else {
+            $process->setStatus(ProcessStatusEnum::ERROR);
+        }
+
+        $this->entityManager->persist($process);
+        if ($flush) {
+            $this->entityManager->flush();
+        }
+    }
+}